GraphiQL replacement => gui middleware option should be playground config instead
Hi,
I don't know if using `gui` was possible in previous version but the way to go seems to be using the `playground` config instead.

Also, passing a header from the localStorage is not possible anymore, like with the `passHeader` option from graphiQL, you would write `graphiqlOptions: {
    passHeader: "'Authorization': localStorage['Meteor.loginToken']", // eslint-disable-line quotes
  },`
Is there a replacement pattern that we could document? Using an auth token from the localStorage is a common pattern.